requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
attrs==26.1.0
bidict==0.23.1
bleach==6.3.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
etl-entities==2.6.0
evacuator==1.0.4
fastapi==0.135.2
frozendict==2.4.7
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.15.0
idna==3.11
importlib_metadata==9.0.0
Jinja2==3.1.6
jiter==0.13.0
json_repair==0.58.7
kaggle==2.0.0
kagglehub==1.0.0
kagglesdk==0.1.16
MarkupSafe==3.0.3
minio==7.2.20
multidict==6.7.1
numpy==2.4.4
onetl==0.15.0
openai==2.30.0
ordered-set==4.1.0
packaging==26.0
pandas==3.0.1
platformdirs==4.9.4
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
psycopg2-binary==2.9.11
py4j==0.10.9.9
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
pyspark==4.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-slugify==8.0.4
PyYAML==6.0.3
requests==2.33.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==1.0.0
text-unidecode==1.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
webencodings==0.5.1
websockets==16.0
yarl==1.23.0
zipp==3.23.0
